tests/wpt/web-platform-tests/trusted-types/Window-TrustedTypes.tentative.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/helper.sub.js"></script>
<body>
<script>
  test(t => {
    let factory = window.TrustedTypes;
    assert_true(factory instanceof TrustedTypePolicyFactory);
  }, "factory = window.TrustedTypes");
</script>
